backend/requirements.txt
flask
gunicorn
google-cloud-firestore
google-cloud-storage
